Fix warning in Spells.cpp: change variable to size_t.

+++ src/game/Spells.cpp
@@ -4679,8 +4679,7 @@ bool ARX_SPELLS_Launch( const long& typ, const long& source, const long& flagss,
 			long valid=0;
 			long dispelled=0;
 
-			for (long n=0;n<MAX_SPELLS;n++)
-			{
+			for(size_t n = 0; n < MAX_SPELLS; n++) {
 				if (!spells[n].exist) continue;
 
 				switch (spells[n].type)
